backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
promise==2.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
